+++ Changelog.md
@@ -10,6 +10,13 @@ Improvements:
 - default to 0 days max age for new databases instead of 365 days
 - add Docusaurus docs
 
+Bug fixes:
+- fix storage not being cleaned when deleting full database
+- fix missing timestamp in JSON inserting timestamp far in the past
+- fix time display for timestamp <1970
+- fix export not including media fields
+- other minor fixes
+
 # v3.0
 
 Features: